pyproject.toml
[project]
name = "cortex-utils"
version = "0.1.0"
description = "Operational utilities for Cortex: queue management, alerting, log tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "isort>=5.13.0",
]

[project.scripts]
cortex-utils = "cortex_utils.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cortex_utils"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
